requirements.txt
# OCR
pytesseract==0.3.10
easyocr==1.7.1
pdfplumber==0.11.0
pymupdf==1.24.3
Pillow==10.3.0

# Layout-aware model
transformers==4.40.2
torch==2.3.0
datasets==2.19.1

# Post-processing / NLP
regex==2024.4.28

# Embeddings + Vector DB
sentence-transformers==3.0.0
faiss-cpu==1.8.0

# MLOps
mlflow==2.13.0

# API + serving
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9

# Utilities
numpy==1.26.4
pandas==2.2.2
pydantic==2.7.1
python-dotenv==1.0.1
tqdm==4.66.4
loguru==0.7.2

# Dev / testing
pytest==8.2.0
pytest-cov==5.0.0
black==24.4.2
ruff==0.4.4
